Fixed linkable dtor by relinking downlinks from backlinks
They get relinked to the first of the backlinks which becomes the new root.   This only matters if the link being deleted was open, had no ref.  Otherwise downlinks are never followed anyway.   When unlinked, we just want to make sure the backlinks continue to be tied together to one root with one eventual possible data link.